+++ common/cards/hermits/smallishbeans-rare.ts
@@ -3,10 +3,12 @@ import query from '../../components/query'
 import {GameModel} from '../../models/game-model'
 import {beforeAttack} from '../../types/priorities'
 import {hermit} from '../defaults'
-import KingJoelCommon from '../hermits/kingjoel-common'
-import KingJoelRare from '../hermits/kingjoel-rare'
 import {Hermit} from '../types'
-import SmallishbeansCommon from './smallishbeans-common'
+import EthosLabCommon from './ethoslab-common'
+import EthosLabRare from './ethoslab-rare'
+import EthosLabUltraRare from './ethoslab-ultra-rare'
+import GeminiTayCommon from './geminitay-common'
+import GeminiTayRare from './geminitay-rare'
 
 const SmallishbeansRare: Hermit = {
 	...hermit,
@@ -15,7 +17,7 @@ const SmallishbeansRare: Hermit = {
 	name: 'Joel',
 	expansion: 'season_x',
 	rarity: 'rare',
-	tokens: 1,
+	tokens: 0,
 	type: 'explorer',
 	health: 260,
 	primary: {
@@ -29,7 +31,7 @@ const SmallishbeansRare: Hermit = {
 		cost: ['explorer', 'explorer', 'any'],
 		damage: 90,
 		power:
-			'For each AFK Joel or King Joel on the game board, do an additional 10hp damage.',
+			'For each your AFK Ethoslab or Geminitay cards on the game board, do an additional 10hp damage.',
 	},
 	onAttach(
 		game: GameModel,
@@ -43,19 +45,21 @@ const SmallishbeansRare: Hermit = {
 				if (!attack.isAttacker(component.entity) || attack.type !== 'secondary')
 					return
 
-				const joelQuantity = game.components.filter(
+				const obsessionAmount = game.components.filter(
 					CardComponent,
 					query.card.attached,
 					query.card.is(
-						SmallishbeansCommon,
-						SmallishbeansRare,
-						KingJoelCommon,
-						KingJoelRare,
+						EthosLabCommon,
+						EthosLabRare,
+						EthosLabUltraRare,
+						GeminiTayCommon,
+						GeminiTayRare,
 					),
+					query.card.currentPlayer,
 					query.not(query.card.active),
 				).length
 
-				attack.addDamage(component.entity, joelQuantity * 10)
+				attack.addDamage(component.entity, obsessionAmount * 10)
 			},
 		)
 	},
